+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Minor
 * Masonry: Makes Masonry React Async compatible (#227)
 * SegmentedControl: Change flow type of `items` to `React.Node` (#230)
+* Video: Add jsdom browser specific tests (#205)
 
 ### Patch
 * Internal: add better basic test coverage (#231)